DEV-99: Re-add aggregates
