+++ app/dashboard/users/page.tsx
@@ -1,17 +1,23 @@
 import { getCurrentUserProfile } from "@/app/lib/users/helpers";
 
+import UsersTableFilters from "@/app/components/users/filters/users-table-filters";
+import TableFiltersSkeleton from "@/app/components/users/skeletons/filters";
+import TableSkeleton from "@/app/components/users/skeletons/table";
 import UsersTable from "@/app/components/users/table";
 import {
-  SearchParamsSchema,
-  SearchParamsSchemaType,
+	SearchParamsSchema,
+	SearchParamsSchemaType,
 } from "@/app/dashboard/users/schemas";
+import {
+	fetchUserProfiles,
+	fetchUsersAggregates,
+} from "@/app/lib/users/actions";
 import { notFound } from "next/navigation";
-import UsersTableFilters from "@/app/components/users/filters/users-table-filters";
-import { fetchUserProfiles } from "@/app/lib/users/actions";
-import TableSkeleton from "@/app/components/users/skeletons/table";
 import { Suspense } from "react";
 
-export default async function Page(props: { searchParams: Promise<SearchParamsSchemaType> }) {
+export default async function Page(props: {
+	searchParams: Promise<SearchParamsSchemaType>;
+}) {
 	const searchParams = await props.searchParams;
 	const validatedSearchParams = SearchParamsSchema.safeParse(searchParams);
 	if (!validatedSearchParams.success) notFound();
@@ -39,14 +45,26 @@ export default async function Page(props: { searchParams: Promise<SearchParamsSc
 		direction,
 		profileCompletion,
 	});
+	const fetchUsersAggregatesPromise = fetchUsersAggregates({
+		includeAdmins,
+		status,
+		category,
+		query,
+		profileCompletion,
+	});
 
 	return (
 		<div className="container mx-auto min-h-full p-3 md:p-6">
 			<h1 className="mb-2 text-2xl font-bold md:text-3xl">Usuarios</h1>
 			<div className="flex flex-col gap-4 group">
-				<UsersTableFilters />
+				<Suspense fallback={<TableFiltersSkeleton />}>
+					<UsersTableFilters />
+				</Suspense>
 				<Suspense fallback={<TableSkeleton />}>
-					<UsersTable fetchUsersPromise={fetchUsersPromise} />
+					<UsersTable
+						fetchUsersPromise={fetchUsersPromise}
+						fetchUsersAggregatesPromise={fetchUsersAggregatesPromise}
+					/>
 				</Suspense>
 			</div>
 		</div>
